% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{dynamic_A}
\alias{dynamic_A}
\title{Get the adjacency matrix for plotting}
\usage{
dynamic_A(x, factor_count, simple.flag, threshold)
}
\arguments{
\item{x}{input the original estimated loading matrix}

\item{factor_count}{The number of factors to use}

\item{simple.flag}{if True, only eliminate the entries below threshold and make all row sums to be 1; if False, the approach further eliminates the entries of the rows that are very close to threshold value and only leaves the maximum entry of each row}

\item{threshold}{A parameter to eliminate very small entries of the loading matrix}
}
\value{
The new loading matrix with all rows sum to be 1
#'#'@seealso \code{\link{MAR1.projection}}            # Not decided
}
\description{
Get the adjacency matrix for plotting
}
