% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.E}
\alias{gen.E}
\title{Generate Error \eqn{E_t} for Matrix}
\usage{
gen.E(n, p1, p2, cov1.half, cov2.half)
}
\arguments{
\item{n}{length of time}

\item{p1}{p1}

\item{p2}{p2}

\item{cov1.half}{p1 * p1 covariance matrix}

\item{cov2.half}{p2 * p2 covariance matrix}
}
\value{
A tensor of size p1 * p2 * n
}
\description{
For simulation only, generate the error tensor E with given two covariance matrices
}
