% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.E.tensor3}
\alias{gen.E.tensor3}
\title{generate error \eqn{E_t} for tensor}
\usage{
gen.E.tensor3(n, p1, p2, p3, cov1.half, cov2.half, cov3.half)
}
\arguments{
\item{n}{length of time}

\item{p1}{p1}

\item{p2}{p2}

\item{p3}{p3}

\item{cov1.half}{p1 * p1 covariance matrix}

\item{cov2.half}{p2 * p2 covariance matrix}

\item{cov3.half}{p3 * p3 covariance matrix}
}
\value{
A tensor of size p1 * p2 * p3 * n
}
\description{
For simulation only, generate the error tensor E with given three covariance matrices
}
