% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.F}
\alias{gen.F}
\title{generate matrix \eqn{F_t}}
\usage{
gen.F(ar1.coef, n, innovsd)
}
\arguments{
\item{ar1.coef}{r1*r2}

\item{n}{length of time}

\item{innovsd}{sd of factor ar series}
}
\value{
An array of size r1*r2*n
}
\description{
For simulation only, generate a matrix F with AR coefficients of length n
}
