% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.F.tensor3}
\alias{gen.F.tensor3}
\title{generate tensor \eqn{F_t}}
\usage{
gen.F.tensor3(ar1.coef, n, innovsd)
}
\arguments{
\item{ar1.coef}{r1*r2*r3}

\item{n}{length of time}

\item{innovsd}{sd of factor ar series}
}
\value{
An array of size r1*r2*r3*n
}
\description{
For simulation only, generate a tensor F with AR coefficients of length n
}
