% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.Q}
\alias{gen.Q}
\title{generate orthonormal loading Q}
\usage{
gen.Q(p, r)
}
\arguments{
\item{p}{row dimension}

\item{r}{column dimension, should not be larger than p}
}
\value{
An orthonormal matrix Q with size p*r
}
\description{
For simulation only, generate loading matrix Q with size p*r
}
