% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.X}
\alias{gen.X}
\title{generate X for matrix}
\usage{
gen.X(Q1, Q2, bF)
}
\arguments{
\item{Q1}{left loading matrix, p1*r1 orthonormal}

\item{Q2}{right loading matrix, p2*r2 orthonormal}

\item{bF}{factor series, r1*r2*n}
}
\value{
A tensor of size p1*p2*n
}
\description{
For simulation only, generate \eqn{X = bF \times_1 Q_1 \times_2 Q_2}
}
