% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{gen.X.tensor3}
\alias{gen.X.tensor3}
\title{generate X for tensor}
\usage{
gen.X.tensor3(Q1, Q2, Q3, bF)
}
\arguments{
\item{Q1}{loading matrix, p1*r1 orthonormal}

\item{Q2}{loading matrix, p2*r2 orthonormal}

\item{Q3}{loading matrix, p3*r3 orthonormal}

\item{bF}{factor series, r1*r2*r3*n}
}
\value{
A tensor of size p1*p2*p3*n
}
\description{
For simulation only, generate \eqn{X = bF \times_1 Q_1 \times_2 Q_2 \times_3 Q_3}
}
