% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{generate.A}
\alias{generate.A}
\alias{generate}
\title{Generate coefficient matrices}
\usage{
generate.A(dim, R)
}
\arguments{
\item{dim}{an array of dimensions of matrices or tensors}

\item{R}{number of terms}
}
\value{
a list containing coefficient matrices
}
\description{
For test only, to generate coefficient matrices
}
\examples{
A <- generate.A(dim=c(2,2,2), R=1)
xx <- generate(c(m1,m2,m3), T=100)
}
