% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{generate.AR}
\alias{generate.AR}
\title{Generate an AR(p) tensor time series}
\usage{
generate.AR(t, dim, R, P, rho, setting)
}
\arguments{
\item{t}{length of time}

\item{dim}{an array of dimensions of matrices or tensors}

\item{R}{number of terms}

\item{P}{number of orders}

\item{rho}{spectral radius of coefficient matrix \eqn{\Phi}}

\item{setting}{the structure of random error, can be specified as "iid", "mle" and "svd", default value is "iid".}
}
\value{
Tensor-valued time series generated by TenAR model
}
\description{
For simulation only, with given time length \code{t}, dimension \code{dim}, number of terms \code{R}, order \code{P}, spectral radius \code{rho} and \code{setting}, generate an AR(1) tensor time series.
}
\examples{
dim <- c(3,3,3)
xx <- generate.AR(t=500, dim, R=2, P=1, rho=0.5, setting='iid')
}
\seealso{
\code{\link{generate.FM}}
}
