% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{generate}
\alias{generate}
\title{Generate an AR(1) tensor time series with given coefficient matrices}
\usage{
generate(A, t, setting = "iid")
}
\arguments{
\item{t}{length of time}

\item{setting}{the structure of random error, can be specified as "iid", "mle" and "svd", default value is "iid".}

\item{dim}{an array of dimensions of matrices or tensors}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{matrices \eqn{A_1,A_2,\cdots,A_k}}
\item{\code{X}}{AR(1) tensor time series}
}
}
\description{
For test only, with given coefficient matrices and time length t, generate an AR(1) tensor time series.
}
\examples{
A <- generate.A(dim=c(2,2,2), R=1)
xx <- generate(c(m1,m2,m3), T=100)
}
\seealso{
\code{\link{run.test}}
}
