% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{get.hat}
\alias{get.hat}
\title{get fitted x.hat(signal part)}
\usage{
get.hat(x, Q, d.seq)
}
\arguments{
\item{x}{tensor of any dimension, x: \eqn{d1 * d2 * d3 * \cdots * d_K * T}}

\item{Q}{list of k matrices}

\item{d.seq}{1:k}
}
\value{
a tensor object x.hat
}
\description{
get x.hat from a tensor x and all loading matrices in list Q by projection
}
