% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{grouping.loading}
\alias{grouping.loading}
\title{Get the group of loadings}
\usage{
grouping.loading(loading, ncluster, rowname, plot = T)
}
\arguments{
\item{loading}{The estimated loading matrix}

\item{ncluster}{The number of clusters to use, usually the dimension of the factor matrix}

\item{rowname}{The name of the rows}

\item{plot}{plot the clustering graph, defacult True}
}
\value{
Loading matrix after grouping
#'#'@seealso \code{\link{MAR1.projection}}            # Not decided
}
\description{
Get the group of loadings
}
