% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{iter.tensor.bic}
\alias{iter.tensor.bic}
\title{iterative versions of all three methods for any Dim tensor time series}
\usage{
iter.tensor.bic(
  x,
  r,
  h0 = 1,
  method,
  tol = 1e-04,
  niter = 100,
  tracetrue = FALSE
)
}
\arguments{
\item{x}{tensor of any dimension, \eqn{x: d1 * d2 * d3 * \cdots * d_K * n}}

\item{r}{initial guess of # of factors}

\item{h0}{Pre-scribed parameter h}

\item{method}{the method chosen among UP,TIPUP,TOPUP}

\item{tol}{level of error tolerance}

\item{niter}{number of iterations}

\item{tracetrue}{if TRUE, record the dis value for each iteration}
}
\value{
a list containing the following:\describe{
\item{\code{Q}}
\item{\code{Qinit}}{initial value of Q}
\item{\code{Qfirst}}{value of Q at first iteration}
\item{\code{x.hat}}
\item{\code{x.hat.init}}{initial value of xhat}
\item{\code{x.hat.first}}{value of xhat at first iteration}
\item{\code{factor.num}}{number of factors}
\item{\code{timer}}{a timer}
\item{\code{norm.percent}}{x after standardization}
\item{\code{dis}}{difference of fnorm.resid for each iteration}
\item{\code{niter}}{number of interations}
\item{\code{fnorm.resid}}{normalized residual}
}
}
\description{
iterative versions of all three methods for any Dim tensor time series
}
