% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR}
\alias{MAR}
\title{Estimation for Autoregressive Model of Matrix-Valued Time Series}
\usage{
MAR(xx, method)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{method}{character string, specifying the method of the estimation to be used. \describe{
 \item{\code{"PROJ",}}{Projection method.}
 \item{\code{"LSE",}}{Iterated least squares.}
 \item{\code{"MLE",}}{MLE under a Kronecker structured covariance matrix.}
 \item{\code{"VAR",}}{Stacked vector AR(1) Model.}
}}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
}
}
\description{
Although MAR model is a special case in tenAR model. We still include this original function as reference.
Estimation function for matrix-valued time series, including the projection method (PROJ),
the Iterated least squares method (LSE) and MLE under a Kronecker structured covariance (MLE) and vectorized AR(1) model (VAR) as determined by the value of \code{method}.
}
\examples{
dim <- c(3,3,3)
A <- tenAR.A(dim,R=2,P=1,rho)
xx <- tenAR.xx(t=500, A, setting='iid')
MAR(xx, method="LSE")
}
