% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.r
\name{MAR.RR}
\alias{MAR.RR}
\title{Reduced Rank MAR(1) iterative estimation for matrix-valued time series}
\usage{
MAR.RR(xx, method=c("RRLSE","RRMLE"), LL.init=NULL, RR.init=NULL,Sigl.init=NULL,
Sigr.init=NULL,k1=NULL, k2=NULL, niter=100,tol=1e-6)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times d_2} matrix-valued time series, \eqn{T} is the length of the series.}

\item{method}{character string, specifying the method of the estimation to be used. \describe{
 \item{\code{"RRLSE",}}{Least squares.}
 \item{\code{"RRMLE",}}{MLE under a separable cov(vec(\eqn{E_t})).}
}}

\item{Sigl.init}{only if \code{method=RRMLE}, initial value of Sigl. The default is the identity matrix.}

\item{Sigr.init}{only if \code{method=RRMLE}, initial value of Sigr. The default is the identity matrix.}

\item{k1}{rank of \eqn{A_1}, a positive integer.}

\item{k2}{rank of \eqn{A_2}, a positive integer.}

\item{niter}{maximum number of iterations if error stays above \code{tol}.}

\item{tol}{relative Frobenius norm error tolerance.}

\item{A1.init}{initial value of \eqn{A_1}. The default is the identity matrix.}

\item{A2.init}{initial value of \eqn{A_2}. The default is the identity matrix.}
}
\value{
return a list containing the following:\describe{
\item{\code{A1}}{estimator of \eqn{A_1}, a \eqn{d_1} by \eqn{d_1} matrix}
\item{\code{A2}}{estimator of \eqn{A_2}, a \eqn{d_2} by \eqn{d_2} matrix}
\item{\code{Sigl}}{only if \code{method=MLE}, when \eqn{\mathrm{Cov}(\mathrm{vec}(E_t))=\Sigma_r \otimes \Sigma_l}.}
\item{\code{Sigr}}{only if \code{method=MLE}, when \eqn{\mathrm{Cov}(\mathrm{vec}(E_t))=\Sigma_r \otimes \Sigma_l}.}
\item{\code{res}}{residuals.}
\item{\code{Sig}}{covariance matrix cov(vec(\eqn{E_t}))}
\item{\code{cov}}{covariance matrix \eqn{\hat A_1} and \eqn{\hat A_2}. If \code{method=RRLSE} or \code{method=RRMLE}, then it is a list containing \describe{
 \item{\code{Sigma.LS}}{asymptotic covariance matrix of (vec( \eqn{\hat A_1}),vec(\eqn{\hat A_2^T})).}
 \item{\code{Theta1.LS.u}, \code{Theta1.LS.v}}{asymptotic covariance matrix of vec(\eqn{\hat U_1}), vec(\eqn{\hat V_1}).}
 \item{\code{Theta2.LS.u}, \code{Theta2.LS.v}}{asymptotic covariance matrix of vec(\eqn{\hat U_2}), vec(\eqn{\hat V_2}).}
}}
\item{\code{sd}}{standard errors of \eqn{\hat A_1} and \eqn{\hat A_2}, returned in a list aligned with \eqn{\hat A_1} and \eqn{\hat A_2}.}
\item{\code{niter}}{number of iterations.}
\item{\code{BIC}}{value of the extended Bayesian information criterion.}
}
}
\description{
Reduced Rank MAR(1) iterative estimation for matrix-valued time series,
including the Reduced Rank MAR(1) iterative estimation (RRLSE), and Reduced Rank MAR(1) iterative estimation with Kronecker covariance structure (RRMLE),
which should be specified by the value of \code{method}.
}
\details{
Matrix autoregressive model (order one) has the following form:
\deqn{X_t = \sum_{r=1}^R A_1^{(r)} X_{t-1} A_2^{(r) \prime} + E_t,}
where \eqn{A_i^{(r)}} are \eqn{d_i \times d_i} coefficient matrices, \eqn{i=1,2}, and \eqn{E_t} is a matrix white noise. \eqn{R} is the number of terms.
Note that only the lag-1 term \eqn{X_{t-1}} appears on the right hand side, so we refer to it as an order-1 model. For order-p model we have the form:
\deqn{X_t = \sum_{i=1}^{p} \sum_{r=1}^{R_i} A_{1}^{(ir)} X_{t-i} A_{2}^{(ir)\prime} + E_t.}
We also consider a special form of Cov(vec(\eqn{E_t})),
 \deqn{\mathrm{Cov}(\mathrm{vec}(E_t))=\Sigma_r \otimes \Sigma_l}
For matrix reduced-rank model, we assume that rank\eqn{(A_i) = k_i \le d_i} for \eqn{i=1,2}.
}
\examples{
dim <- c(3,3,3)
xx <- TenAR.sim(t=500, dim, R=2, P=1, rho=0.5, cov='iid')
est <- MAR.RR(xx, method="RRLSE", k1=1, k2=1)
}
