% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR.SE}
\alias{MAR.SE}
\title{Asymptotic Covariance Matrix of \code{MAR1.MLE}}
\usage{
MAR.SE(xx, B, A, Sigma)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{B}{q by q matrix in MAR(1) model}

\item{A}{p by p matrix in MAR(1) model}

\item{Sig}{covariance matrix cov(vec(E_t)) in MAR(1) model}
}
\value{
asymptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of \code{MAR1.MLE} for given A, B and matrix-valued time series xx, see Theory 3 in paper.
}
\examples{
# given T * p * q time series xx
out = MAR1.LS(xx)
FnormLL = sqrt(sum(out$LL))
xdim=3;ydim=4
out.Xi=MAR.SE(xx.nm,out$RR*FnormLL,out$LL/FnormLL,out$Sig)
out.SE=sqrt(diag(out.Xi))
SE.A=matrix(out.SE[1:xdim^2],nrow=xdim)
SE.B=t(matrix(out.SE[-(1:xdim^2)],nrow=ydim))
}
