% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR1.CC}
\alias{MAR1.CC}
\title{Reduced Rank MAR(1) iterative estimation with Kronecker covariance structure}
\usage{
MAR1.CC(xx, k1, k2, LL.init=NULL,RR.init=LL,Sigl.init=NULL,Sigr.init=NULL,
niter=200,tol=1e-4,print.true = FALSE)
}
\arguments{
\item{xx}{matrix-valued time series}

\item{k1}{rank of A1}

\item{k2}{rank of A2}

\item{LL.init}{initial values of A1 in iterations, by default is diagonal matrices}

\item{RR.init}{initial values of A2 in iterations, by default is diagonal matrices}

\item{Sigl.init}{initial values of Sigma_1 in iterations, by default is diagonal matrices}

\item{Sigr.init}{initial values of Sigma_2 in iterations, by default is diagonal matrices}

\item{niter}{maximum number of iterations if error stays above \code{tol}, by default niter=200}

\item{tol}{relative Frobenius norm error tolerance, by default tol=1e-4}

\item{print.true}{print \eqn{A_i}}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimated \eqn{A_1}}
\item{\code{RR}}{estimated \eqn{A_2}}
\item{\code{Sigl}}{estimated Sigma_1}
\item{\code{Sigr}}{estimated Sigma_2}
\item{\code{res}}{residual}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{dis}}{Frobenius norm difference of last update}
\item{\code{niter}}{number of iterations}
}
}
\description{
MLE for the matrix reduced rank model \eqn{X_t = LL \times X_{t-1} \times {RR}^{\prime} + E_t}.
}
\examples{
dim <- c(4,5) # dimension of matrix at time t is 4 * 5
A <- tenAR.A(dim, R=1, P=1, rho=0.5)
xx <- tenAR.xx(t=500, A, setting="iid")
model <- MAR1.RR(xx, k1=2, k2=2)
}
