% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR1.LS}
\alias{MAR1.LS}
\title{Least Squares Iterative Estimation for Matrix Time Series}
\usage{
MAR1.LS(xx, niter = 50, tol = 1e-06, print.true = FALSE)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{printe LL and RR}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{dis}}{Frobenius norm difference of last update}
\item{\code{niter}}{number of iterations}
}
}
\description{
Iterated least squares estimation in the model \eqn{X_t = LL \times X_{t-1} \times {RR}^{\prime} + E_t}.
}
