% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{MAR1.otimes}
\alias{MAR1.otimes}
\title{MLE under a structured covariance tensor}
\usage{
MAR1.otimes(
  xx,
  LL.init = NULL,
  Sigl.init = NULL,
  Sigr.init = NULL,
  niter = 50,
  tol = 1e-06,
  print.true = FALSE
)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{LL.init}{initial value of LL}

\item{Sigl.init}{initial value of Sigl}

\item{Sigr.init}{initial value of Sigr}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{printe LL and RR}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sigl}}{one part of structured covariance matrix Sig=Sigr otimes Sigl}
\item{\code{Sigr}}{one part of structured covariance matrix Sig=Sigr otimes Sigl}
\item{\code{dis}}{Frobenius norm difference of last update}
\item{\code{niter}}{number of iterations}
}
}
\description{
MAR(1) iterative estimation with Kronecker covariance structure: \eqn{X_t = LL X_{t-1} RR + E_t} such that \eqn{Sig = cov(vec(E_t)) = Sigr otimes Sigl}.
}
