% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR1.PROJ}
\alias{MAR1.PROJ}
\title{Projection Method for MAR(1)}
\usage{
MAR1.PROJ(xx)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
}
}
\description{
MAR(1) one step projection estimation in the model \eqn{X_t = LL \times X_{t-1} \times {RR}^{\prime} + E_t}.
}
