% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR1.RR}
\alias{MAR1.RR}
\title{Reduced Rank MAR(1) iterative estimation}
\usage{
MAR1.RR(xx, k1, k2, niter=200, tol=1e-4, print.true=FALSE,
LL.init=NULL, RR.init=NULL)
}
\arguments{
\item{xx}{matrix-valued time series}

\item{k1}{rank of A1}

\item{k2}{rank of A2}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{print \eqn{A_i}}

\item{LL.init}{initial values of A1 in iterations, by default is diagonal matrices}

\item{RR.init}{initial values of A2 in iterations, by default is diagonal matrices}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of coefficient matrices \eqn{A_1}}
\item{\code{RR}}{estimator of coefficient matrices \eqn{A_2}}
\item{\code{res}}{residual}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{BIC}}{value of information criteria}
\item{\code{dis}}{Frobenius norm difference of last update}
\item{\code{niter}}{number of iterations}
}
}
\description{
LSE for the matrix reduced rank model \eqn{X_t = LL \times X_{t-1} \times {RR}^{\prime} + E_t}.
}
\examples{
dim <- c(4,5) # dimension of matrix at time t is 4 * 5
A <- tenAR.A(dim, R=1, P=1, rho=0.5)
xx <- tenAR.xx(t=500, A, setting="iid")
model <- MAR1.RR(xx, k1=2, k2=2)
}
