% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{MAR2.otimes}
\alias{MAR2.otimes}
\alias{MAR1.otimes}
\title{MLE under a structured covariance tensor for Matrix Time Series with Multiple Terms}
\usage{
MAR2.otimes(xx, r, niter = 200, tol = 1e-06, print.true = FALSE)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{print LL and RR}

\item{LL.init}{initial value of LL}

\item{Sigl.init}{initial value of Sigl}

\item{Sigr.init}{initial value of Sigr}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{SIGMA}}{structured covariance matrix \eqn{\Sigma=\Sigma_1, \cdots, \Sigma_R}}
\item{\code{dis}}{Frobenius norm difference of the final update step}
\item{\code{niter}}{number of iterations}
}
}
\description{
MAR(1) iterative estimation with Kronecker covariance structure: \eqn{X_t = \sum_{r=1}^{R} LL_r * X_{t-1} * RR_r + E_t} such that \eqn{\Sigma = cov(vec(E_t)) = \Sigma_r \otimes \Sigma_l}.
}
