% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{MAR2.projection}
\alias{MAR2.projection}
\title{Projection Method for matrix time series with multiple terms}
\usage{
MAR2.projection(xx)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}
}
\value{
a list containing the estimated coefficient matrices
}
\description{
MAR(1) one step projection estimation in the model \eqn{X_t = \sum_{r=1}^{R} LL_r * X_{t-1} * RR_r + E_t}.
}
