% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{matrix_factor}
\alias{matrix_factor}
\title{matrix_factor}
\usage{
matrix_factor(Yt, inputk1, inputk2, iscentering = 1, hzero = 1)
}
\arguments{
\item{Yt}{Time Series data for a matrix}

\item{inputk1}{The pre-determined row dimension of the factor matrix}

\item{inputk2}{The pre-determined column dimension of the factor matrix}

\item{iscentering}{The data is subtracted by its mean value}

\item{hzero}{Pre-determined parameter}
}
\value{
a list containing the following:\describe{
\item{\code{eigval1}}{estimated row dimension of the factor matrix}
\item{\code{eigval2}}{estimated column dimension of the factor matrix}
\item{\code{loading1}}{estimated left loading matrix}
\item{\code{loading2}}{estimated right loading matrix}
\item{\code{Ft}}{Estimated factor matrix with pre-determined number of dimensions}
\item{\code{Ft.all}}{Sum of Ft}
\item{\code{Et}}{The estimated residual, by subtracting estimated signal term from the data}
}
}
\description{
The main estimation function
}
\examples{
A <- 1:180
dim(A) <- c(3,3,20)
out = matrix_factor(A,3,3)
eig1 = out$eigval1
loading1 = out$loading1
Ft = out$Ft.all
}
