% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda.estqk}
\alias{mfmda.estqk}
\title{Compute the estimated number of factors and the corresponding eigen-space}
\usage{
mfmda.estqk(Mhat, inputk = 1)
}
\arguments{
\item{Mhat}{The estimated value for matrix M}

\item{inputk}{The pre-determined number of dimension of factor matrix}
}
\value{
The estimated number of factors to use, the corresponding estimated Q matrix, the eigenvalue,  the estimated Q matrix with requested number of factors
#'#'@seealso \code{\link{MAR1.projection}}            # Not decided
}
\description{
Compute the estimated number of factors and the corresponding eigen-space
}
\examples{
A <- 1:180
dim(A) <- c(3,3,20)
M <- mfmda(A,"3",1,0)
inputk <- 3
eig.ans <- mfmda.estqk(M,inputk)
khat <- eig.ans$estk
Qhat <- eig.ans$Qhatestk
eigval <- eig.ans$eigval
Qhatinputk <- eig.ans$Qhatinputk
}
