% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda.nona.iter}
\alias{mfmda.nona.iter}
\title{mfmda.nona.iter}
\usage{
mfmda.nona.iter(Yc, hzero)
}
\arguments{
\item{Yc}{Time Series data for a matrix(dimensions n*p*q), no NA input allowed}

\item{hzero}{Pre-scribed parameter}
}
\value{
The sample version of M matrix
}
\description{
The input data do not have zeros. The estimation approach is iterative.
}
