% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda.nona.vec}
\alias{mfmda.nona.vec}
\title{This approach is for the vector-valued estimation WITHOUT NaNs.}
\usage{
mfmda.nona.vec(Yc, hzero)
}
\arguments{
\item{Yc}{Time Series data for a matrix(dimensions n*p*q), no NA input allowed}

\item{hzero}{Pre-scribed parameter}
}
\value{
The sample version of M matrix
#'#'@seealso \code{\link{MAR1.projection}}            # Not decided
}
\description{
This approach is for the vector-valued estimation WITHOUT NaNs.
}
\examples{
A <- 1:180
dim(A) <- c(3,3,20)
M <- mfmda.nona.vec(A,2)
}
