% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda}
\alias{mfmda}
\title{This is a wrapper for all approaches}
\usage{
mfmda(Yt, approach = "3", hzero = 1, iscentering = 1)
}
\arguments{
\item{approach}{Select estimation approaches, 1 for noniterative approach with no NaNs, 2 for iterative approach with NaNs, 3 for iterative approach allowing NaNs.}

\item{hzero}{Pre-determined parameter}

\item{iscentering}{The data is subtracted by its mean value}

\item{Yc}{Time Series data for a matrix}
}
\value{
The sample version of M matrix
}
\description{
This is a wrapper for all approaches
}
\examples{
#Use iterative method with $h_0=1$
A <- runif(2000)
dim(A) <- c(20,10,10)
M <- mfmda(A,"3",1,0)
}
\seealso{
\code{\link{matrix_factor}}
}
