% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{mplot.acf}
\alias{mplot.acf}
\title{Plot ACF of Matrix-Valued Time Series}
\usage{
mplot.acf(xx)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times d_2} matrix-valued time series. Note that the number of mode is 3, where the first mode is time.}
}
\value{
a figure.
}
\description{
Plot ACF of matrix-valued time series, can be also used to plot ACF of a given slice of a tensor-valued time series.
}
\examples{
dim <- c(3,3,3)
xx <- tenAR.sim(t=50, dim, R=2, P=1, rho=0.5, cov='iid')
mplot.acf(xx[1:30,,,1])
}
