% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{PlotNetwork_AB}
\alias{PlotNetwork_AB}
\title{Plot the network graph}
\usage{
PlotNetwork_AB(Ft, iterated_A, iterated_B, labels)
}
\arguments{
\item{Ft}{The estimated factor matrix}

\item{iterated_A}{The left loading matrix}

\item{iterated_B}{The right loading matrix}

\item{labels}{The row labels}
}
\value{
Plot the network graph
}
\description{
Plot the network graph
}
\examples{
A <- runif(2000)
dim(A) <- c(20,10,10)
out = matrix_factor(A,3,3)
eig1 = out$eigval1
loading1 = out$loading1
loading2 = out$loading2
Ft = out$Ft.all
iterated_A = dynamic_A(loading1,3,F,0.1)
iterated_B = dynamic_A(loading2,3,F,0.1)
labels = 1:10
PlotNetwork_AB(Ft,iterated_A,iterated_A,labels)
}
