% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{pm}
\alias{pm}
\title{Permutation matrix pm}
\usage{
pm(m, n)
}
\arguments{
\item{m}{an array of dimentions of matrices \eqn{A_1,A_2,\cdots,A_k}}

\item{n}{length of time}
}
\value{
Permutation matrix pm
#'@seealso \code{\link{trearrange}}
}
\description{
Permutation matrix pm.
}
\examples{
pm(m,n)
}
