% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{predict.rolling}
\alias{predict.rolling}
\title{Model Predictions by rolling forecast}
\usage{
predict.rolling(object, data, n.head, method="LSE")
}
\arguments{
\item{object}{a fit from TenAR(P) model}

\item{data}{data to which to apply the prediction.}

\item{n.head}{number of steps ahead at which to predict.}

\item{method}{method used by rolling forecast}
}
\value{
predicted value
}
\description{
predict.rolling is a function for predictions from the results of model fitting functions.
The function invokes particular methods which depend on the class of the first argument.
Our function is similar to the usage of classical `predict.ar` in package "stats".
}
\examples{
dim <- c(2,2,2)
A <- TenAR.A(dim,R=2,P=1,rho)
xx <- TenAR.xx(t=500, A, setting='iid')
pred.xx <- predict.rolling(model, xx, n.head = 5)
}
\seealso{
`predict.ar` or `predict.arima`
}
