% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{predict.tenar}
\alias{predict.tenar}
\title{Model Predictions}
\usage{
\method{predict}{tenar}(object, data, n.head, se.fit = TRUE, roling = FALSE, method = "LSE")
}
\arguments{
\item{object}{a fit from TenAR()}

\item{data}{data to which to apply the prediction.}

\item{n.head}{number of steps ahead at which to predict.}

\item{se.fit}{logical: return estimated standard errors of the prediction error. default is TRUE.}

\item{method}{method used by rolling forecast}

\item{rolling}{logical: use rolling forcast}
}
\value{
predicted value

Our function is similar to the usage of classical `predict.ar` in package "stats".
}
\description{
predict.tenar is a function for predictions from the results of model fitting functions. The function invokes particular methods which depend on the class of the first argument.
}
\examples{
dim <- c(2,2,2)
A <- generateA(dim, R=1)
xx <- generate(dim, T=100)
pred.xx <- predict.tenar(model, xx, n.head = 5)
}
\seealso{
`predict.ar` or `predict.arima`
}
