% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{Rank.est}
\alias{Rank.est}
\title{Tensor factor rank determination}
\usage{
Rank.est(x,r,rank='BIC',method='TIPUP')
}
\arguments{
\item{x}{\eqn{m_1 * \cdots * m_K * T} tensor-valued time series}

\item{r}{input rank of factor matrix}

\item{h0}{lag number}

\item{rank}{character string, specifying the type of the rank determination method to be used. \describe{
 \item{\code{"BIC",}}{Bayesian information criterion}
 \item{\code{"ER",}}{Eigen ratio method}
}}

\item{method}{character string, specifying the type of the factor estimation method to be used. \describe{
 \item{\code{"TIPUP",}}{TIPUP method}
 \item{\code{"TOPUP",}}{TOPUP method}
}}

\item{inputr}{boolean, if TRUE, use input rank for each iteration; if FLASE, update the rank r in each iteration}

\item{iter}{boolean, specifying using an iterative approach or an non-iterative approach}

\item{penalty}{takes value in {1,2,3,4,5}, decide which penalty function to use}

\item{delta1}{weakest factor strength, a tuning parameter used for BIC method only}

\item{tol}{tolerance for each iteration}

\item{maxiter}{integer, the max number of iterations allowed}
}
\value{
a list containing the following:\describe{
\item{\code{factor.num}}{estimated number of factors \eqn{n_1,n_2,\cdots,n_K}}
}
}
\description{
Function for rank determination of the factor
}
\examples{
Q1 = gen.Q(5,2)
Q2 = gen.Q(5,2)
ar_coef = array(c(0.3,0.5,0.4,0.6),c(2,2))
bF = gen.F(ar_coef,20,1)
x = aperm(gen.X(Q1,Q1,bF),c(3,1,2))
x = as.tensor(x) # Generate x and permute the axes
rank <- Rank.est(x,c(2,2),rank='BIC',method='TIPUP')  # Estimate the rank
}
