% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{rearrange}
\alias{rearrange}
\title{Rearrangement Operator}
\usage{
rearrange(A, m1, m2, n1, n2)
}
\arguments{
\item{A}{m by n matrix such that \eqn{m = m1*n1} and \eqn{n = m2*n2}}

\item{m1}{\code{ncol} of A}

\item{m2}{\code{ncol} of B}

\item{n1}{\code{nrow} of A}

\item{n2}{\code{nrow} of B}
}
\value{
rearengement matrix
#'@seealso \code{\link{MAR1.projection}}
}
\description{
Rearrangement Operator used for projection method.
}
\examples{
A <- matrix(runif(6),ncol=2),
B <- matrix(runif(6),ncol=2),
M <- kronecker(B,A)
rearrange(M,3,3,2,2) == t(as.vector(A)) \%*\% as.vector(B)
'TRUE'
}
