% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR}
\alias{TAR}
\title{Tensor Method}
\usage{
TAR(xx, type = c("projection", "LS", "MLE", "ar"))
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{method}{character string, specifying the type of the estimation method to be used. \describe{
 \item{\code{"projection",}}{Projection method.}
 \item{\code{"lse",}}{Iterated least squares.}
 \item{\code{"mle",}}{MLE under a structured covariance tensor.}
 \item{\code{"ar",}}{Stacked vector AR(1) Model.}
}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coeficient matrices \eqn{A_1,A_2,\cdots,A_K}}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
Estimation function for tensor-valued time series.
Projection method, the Iterated least squares method, MLE under a structured covariance tensor and stacked vector AR(1) model, as determined by the value of \code{type}.
}
