% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR.SE}
\alias{TAR.SE}
\title{Asymptotic Covariance Matrix of \code{TAR1.LS}}
\usage{
TAR.SE(xx, C, B, A, Sigma)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{C}{m3 by m3 matrix in MAR(1) model}

\item{B}{m2 by m2 matrix in MAR(1) model}

\item{A}{m1 by m1 matrix in MAR(1) model}

\item{Sig}{covariance matrix cov(vec(E_t)) in TAR(1) model}
}
\value{
asmptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of \code{TAR1.LS} for given A tensor-valued time series xx, see Theory 2 in paper.
}
