% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR1.projection}
\alias{TAR1.projection}
\title{Projection Method for Tensor-Valued Time Series}
\usage{
TAR1.projection(xx)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{m1}{dim(A1)}

\item{m2}{dim(A2)}

\item{m3}{dim(A3)}
}
\value{
a list containing the estimation of matrices \eqn{A_1,A_2,\cdots,A_K}
}
\description{
TenAR(1) one step projection estimation in the model \eqn{X_t = X_{t-1} \times A_1 \times \cdots \times A_K + E_t}.
}
