% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR1.VAR}
\alias{TAR1.VAR}
\title{Stacked vector AR(1) Model for Tensor-Valued Time Series}
\usage{
TAR1.VAR(xx)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}
}
\value{
a list containing the following:\describe{
\item{\code{coef}}{coeficient of the fitted VAR(1) model}
\item{\code{res}}{residual of the VAR(1) model}
}
}
\description{
vector AR(1) Model.
}
\examples{

}
