% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR2.LS}
\alias{TAR2.LS}
\title{Least Squares Iterative Estimation for Tensor-Valued Time Series with Multiple Terms}
\usage{
TAR2.LS(xx, r, niter = 80, tol = 1e-06, print.true = FALSE)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{print \eqn{A_i}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coefficient matrices \eqn{A_1^{(1)},A_2^{(1)},\cdots,A_K^{(R)}}}
\item{\code{res}}{residual of the model}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
Iterated least squares estimation in the model \eqn{X_t = \sum_{r=1}^{R} X_{t-1} \times A_1^{(r)} \times \cdots \times A_K^{(r)} + E_t}.
}
