% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR2.MLE}
\alias{TAR2.MLE}
\alias{TAR1.MLE}
\title{MLE for Tensor-Valued Time Series with Multiple Terms Model Under a Structured Covariance Tensor}
\usage{
TAR2.MLE(xx, r, niter = 200, tol = 1e-06, print.true = FALSE)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{printe \eqn{A_i}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coeficient matrices \eqn{A_1,A_2,\cdots,A_K}}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
MLE for the model \eqn{X_t = \sum_{r=1}^{R} X_{t-1} \times A_1^{(r)} \times \cdots \times A_K^{(r)} + E_t}.
}
