% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR2.projection}
\alias{TAR2.projection}
\title{Projection Method for Tensor-Valued Time Series with Multiple Terms}
\usage{
TAR2.projection(xx, r)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{r}{number of terms}
}
\value{
a list containing the estimation of matrices \eqn{A_1^{(1)},A_2^{(1)},\cdots,A_K^{(R)}}
}
\description{
TAR(1) one step projection estimation in the model \eqn{X_t = \sum_{r=1}^{R} X_{t-1} \times A_1^{(r)} \times \cdots \times A_K^{(r)} + E_t}.
}
