% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR2.SE.MLE}
\alias{TAR2.SE.MLE}
\title{Asymptotic Covariance Matrix of \code{TAR2.MLE}}
\usage{
TAR2.SE.MLE(xx, A.true, Sigma)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{A.true}{coefficient matrices in TAR(1) model}

\item{Sigma}{covariance matrix cov(vec(E_t)) in TAR(1) model}
}
\value{
asmptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of \code{TAR2.MLE} for given A tensor-valued time series xx, see related Theorems in our paper.
}
\examples{
dim <- c(2,2,2)
A <- generateA(dim, R=1)
xx <- generate(dim, T=100)
SIGMA <- TAR2.SE.MLE(xx, A, Sigma=diag(prod(dim)))
}
