% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{tenAR.A}
\alias{tenAR.A}
\title{Generate coefficient matrices}
\usage{
tenAR.A(dim, R, P, rho)
}
\arguments{
\item{dim}{an array of dimensions of matrices or tensors}

\item{R}{number of terms}

\item{P}{number of orders}

\item{rho}{spectral radius}
}
\value{
a list containing coefficient matrices
}
\description{
For test only, generate coefficient matrices in TenAR model
}
\examples{
dim <- c(3,3,3)
A <- tenAR.A(dim,R=2,P=1,rho=0.5)
xx <- tenAR.xx(t=500, A, setting='iid')
}
