% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{tenAR.LS}
\alias{tenAR.LS}
\title{Least Squares Iterative Estimation for TenAR(p)}
\usage{
tenAR.LS(xx, R, P, init = NULL, niter = 500, tol = 1e-06, print.true = FALSE)
}
\arguments{
\item{xx}{\eqn{T * d_1 * \cdots * d_K} tensor-valued time series}

\item{R}{number of terms}

\item{P}{number of orders}

\item{init}{initial value of A}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{print \eqn{A_i}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coefficient matrices}
\item{\code{res}}{residual of the model}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
Iterated least squares estimation
}
