% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.MLE}
\alias{TenAR.MLE}
\title{MLEs for TenAR(p) Model}
\usage{
TenAR.MLE(
  xx,
  R,
  P,
  init.A = NULL,
  init.sig = NULL,
  niter = 500,
  tol = 1e-05,
  print.true = FALSE
)
}
\arguments{
\item{xx}{\eqn{T * d_1 * \cdots * d_K} tensor-valued time series}

\item{R}{number of terms}

\item{P}{number of orders}

\item{init.A}{initial value of A}

\item{init.sig}{initial value of Sigma}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{print \eqn{A_i}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coefficient matrices}
\item{\code{sigma}}{estimator of Sigma}
\item{\code{res}}{residual}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
MLEs under Kronecker structures covariance matrix
}
