% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.r
\name{tenAR.predict}
\alias{tenAR.predict}
\title{Predictions for Tensor Autoregressive Models}
\usage{
tenAR.predict(object, xx, n.head, method, rolling=FALSE)
}
\arguments{
\item{object}{a fit from TenAR(P) model}

\item{n.head}{number of steps ahead at which to predict}

\item{method}{method used by rolling forecast}

\item{rolling}{TRUE or FALSE, rolling forcast, is FALSE by default}

\item{data}{data to which to apply the prediction}
}
\value{
predicted value
}
\description{
`tenAR.predict` is a function for predictions from the results of model fitting functions.
The function invokes particular methods which depend on the class of the first argument.
If \code{rolling=TRUE}, this would be a rolling forecast, specifically, using given model parameters object,
start from the last time point \eqn{X_t} in data, we obtain the one step ahead prediction \eqn{X_{t+1}}.
Then fit the corresponding model using all available data and \eqn{X_{t+1}}
to obtain \eqn{X_{t+2}}. Repeat this process to \eqn{X_{t+n.head}}.
Our function is similar to the usage of classical `predict.ar' in package "stats".
}
\examples{
dim <- c(2,2,2)
xx <- tenAR.sim(t=500, dim,R=2,P=1,rho=0.5, cov='iid')
est <- tenAR.est(xx, R=1, P=1, method="LSE")
pred <- tenAR.predict(est, xx, n.head = 5)
# rolling forcast
pred.rolling <- tenAR.predict(est, xx, n.head = 5, method="LSE", rolling=TRUE)
}
\seealso{
'predict.ar' or 'predict.arima'
}
