% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.r
\name{TenAR.sim}
\alias{TenAR.sim}
\title{Generate an TenAR(p) tensor time series}
\usage{
TenAR.sim(t, dim, R, P, rho, cov)
}
\arguments{
\item{t}{length of output series. A strictly positive integer.}

\item{dim}{dimension of the tensor at each time.}

\item{R}{Kronecker rank for each lag.}

\item{P}{autoregressive order.}

\item{rho}{spectral radius of coefficient matrix \eqn{\Phi}.}

\item{cov}{covariance matrix of the error term: diagonal ("iid"), separable ("mle"), random ("svd").}
}
\value{
A tensor-valued time series generated by TenAR(p) model.
}
\description{
Simulate from the TenAR(p) model.
}
\examples{
dim <- c(3,3,3)
xx <- TenAR.sim(t=500, dim, R=2, P=1, rho=0.5, cov='iid')
}
\seealso{
\code{\link{tenFM.sim}}
}
