% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.VAR}
\alias{TenAR.VAR}
\title{Stacked vector AR(p) Model}
\usage{
TenAR.VAR(xx, P)
}
\arguments{
\item{xx}{\eqn{T * d_1 * \cdots * d_K} tensor-valued time series}

\item{P}{number of orders}
}
\value{
a list containing the following:\describe{
\item{\code{coef}}{coeficient of the fitted VAR(1) model}
\item{\code{res}}{residual of the VAR(1) model}
}
}
\description{
vector AR(p) Model for vectorized Tensor time series
}
\examples{
out.var = TenAR.VAR(xx)
sum(out.var$res**2)
}
