% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.xx}
\alias{TenAR.xx}
\title{Generate an AR(1) tensor time series with given coefficient matrices}
\usage{
TenAR.xx(t, A, setting)
}
\arguments{
\item{t}{length of time}

\item{A}{true coefficient matrices (the dimensions, number of terms and orders are implied by structure of A)}

\item{setting}{the structure of random error, can be specified as "iid", "mle" and "svd", default value is "iid".}
}
\value{
Tensor-valued time series generated by TenAR model
}
\description{
For test only, with given coefficient matrices and time length t, generate an AR(1) tensor time series.
}
\examples{
dim <- c(3,3,3)
A <- TenAR.A(dim,R=2,P=1,rho=0.5)
xx <- TenAR.xx(t=500, A, setting='iid')
}
\seealso{
\code{\link{run.test}}
}
