% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{tenFM.bic}
\alias{tenFM.bic}
\title{BIC estimators for determing the numbers of factors}
\usage{
tenFM.bic(reigen, h0 = 1, p1, p2, n, delta1 = 0)
}
\arguments{
\item{reigen}{list of eigenvalues}

\item{h0}{lag number}

\item{p1}{p1}

\item{p2}{p2}

\item{n}{n}
}
\value{
factor.p1: Estimated number of factors
}
\description{
BIC estimators for determing the numbers of factors
}
