% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{tenFM.sim}
\alias{tenFM.sim}
\title{Generate tensor time series from given factor structure}
\usage{
tenFM.sim(dims, ar1.coef, t)
}
\arguments{
\item{dims}{dimensions of the output tensor time series,  \eqn{d_1\times d_2\cdots\times d_K}}

\item{ar1.coef}{ar coefficients of the factor process, same dimension with \eqn{F_t : r_1\times r_2\cdots\times r_K}}

\item{t}{length of time}
}
\value{
A tensor of size  \eqn{T\times d_1\times d_2\cdots\times d_K}
}
\description{
For simulation only, generate tensor time series \eqn{X_t} from given factor structure, each factor entry follows an univariate AR(1) process
}
\examples{
dims <- c(4,5,6) # dimensions of tensor time series
t <- 50
r <- c(2,2,2)  # dimensions of factor series
ar1.coef = array(c(runif(prod(r)),min=0.5,max=1)),r)
x = tenFM.sim(dims,ar1.coef,t)  # generate t*dims tensor
}
\seealso{
\code{\link{generate.AR}}
}
