% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor.bic}
\alias{tensor.bic}
\title{BIC estimators for determing the numbers of factors}
\usage{
tensor.bic(reigen, h0 = 1, p1, p2, n)
}
\arguments{
\item{reigen:}{list of eigenvalues}

\item{h0:}{Pre-scribed parameter h}

\item{p1:}{p1}

\item{p2:}{p2}
}
\value{
factor.p1: Estimated number of factors
}
\description{
BIC estimators for determing the numbers of factors
}
