% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{topup.init.tensor}
\alias{topup.init.tensor}
\title{TOPUP one step for any Dim tensor time series}
\usage{
topup.init.tensor(x, r, h0 = 1, oneside.true = FALSE, norm.true = FALSE)
}
\arguments{
\item{x}{tensor of any dimension \eqn{d1 * d2 * d3 * \cdots * d_K * T}}

\item{r}{input rank of factor}

\item{h0}{lag number}

\item{oneside.true}{If TRUE, only compute the one sided column space, not the other sides, this option is useful for the iterative method}

\item{norm.true}{If TRUE, calculate the normalized residual of the tensor}
}
\value{
a list containing the following:\describe{
\item{\code{M}}{Estimator}
\item{\code{Q}}{Orthonormal matrix Q}
\item{\code{lambda}}{singular values}
\item{\code{norm.percent}}{Frobenius norm of residual}
\item{\code{x.hat}}{Fitted signal part, a tensor object}
}
}
\description{
TOPUP one step for any Dim tensor time series
}
