% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{up.init.tensor}
\alias{up.init.tensor}
\title{UP one step for any Dim tensor time series}
\usage{
up.init.tensor(x, r, oneside.true = FALSE, norm.true = FALSE)
}
\arguments{
\item{x}{tensor of any dimension : \eqn{d1 * d2 * d3 * \cdots * d_k * n}}

\item{r}{initial guess of # of factors}

\item{oneside.true}{If oneside.true==TRUE, then only compute the one sided column space, not the other sides, this option is useful for the iterative method}

\item{norm.true}{If norm.true==TRUE, normalize the tensor}
}
\value{
a list containing the following:\describe{
\item{\code{Q}}{Orthonormal matrix Q}
\item{\code{lambda}}{Singular values}
\item{\code{norm.percent}}{x after standardization}
\item{\code{x.hat}}{A tensor object x.hat}
}
}
\description{
UP one step for any Dim tensor time series
}
