% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{var1}
\alias{var1}
\title{Stacked vector AR(1) Model}
\usage{
var1(xx)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}
}
\value{
a list containing the following:\describe{
\item{\code{coef}}{coeficient of the fitted VAR(1) model}
\item{\code{res}}{residual of the VAR(1) model}
}
}
\description{
vector AR(1) Model.
}
\examples{
out.var1=var1(xx)
sum(out.var1$res**2)
}
