% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{vector_factor}
\alias{vector_factor}
\title{The main estimation function, vector version}
\usage{
vector_factor(Yt, inputk.vec, iscentering = 1, hzero = 1)
}
\arguments{
\item{Yt}{Time Series data for a matrix}

\item{inputk.vec}{The pre-determined dimensions of the factor matrix in vector}

\item{iscentering}{The data is subtracted by its mean value}

\item{hzero}{Pre-determined parameter h_0}
}
\value{
a list containing the following:\describe{
\item{\code{eigval1}}{estimated dimensions of the factor matrix}
\item{\code{loading}}{estimated loading matrices}
\item{\code{Ft}}{Estimated factor matrix with pre-determined number of dimensions}
\item{\code{Ft.all}}{Sum of Ft}
\item{\code{Et}}{The estimated random term, by subtracting estimated signal term from the data}
}
}
\description{
The main estimation function, vector version
}
