% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{vts}
\alias{vts}
\title{get M matrix and first r eigenvectors of M}
\usage{
vts(x, h0, r)
}
\arguments{
\item{x}{a n*d matrix}

\item{h0}{lag number}

\item{r}{First r eigenvectors}
}
\value{
a list containing the following:\describe{
\item{\code{M}}{.}
\item{\code{Q}}{The first r eigenvectors of matrix M}
}
}
\description{
get M matrix and first r eigenvectors of M
}
