% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{generate}
\alias{generate}
\title{Generate an AR(1) tensor time series}
\usage{
generate(dim, t)
}
\arguments{
\item{dim}{an array of dimentions of matrices \eqn{A_1,A_2,\cdots,A_k}}

\item{t}{length of time}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{matrices \eqn{A_1,A_2,\cdots,A_k}}
\item{\code{X}}{AR(1) tensor time series}
}
#'@seealso \code{\link{run.test}}
}
\description{
For test only, randomly generate matrices A1,A2,..Ak by iid standard gaussians with given dimensions, and then generate an AR(1) tensor time series.
}
\examples{
dim <- c(1,2,3)
T <- 100
xx <- generate(c(m1,m2,m3),T)
}
