% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{get.hat}
\alias{get.hat}
\title{functions for algorithm}
\usage{
get.hat(x, Q, d.seq)
}
\arguments{
\item{x}{tensor of any dimension, x: \eqn{d1 * d2 * d3 * \cdots * d_k * n}}

\item{Q}{list of k matrices}

\item{d.seq}{1:k}
}
\value{
a tensor object x.hat
}
\description{
functions for algorithm
}
