% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{MAR}
\alias{MAR}
\title{Matrix Method}
\usage{
MAR(xx, type = c("projection", "LS", "MLE", "ar"))
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{type}{character string, specifying the type of the estimation method to be used. \describe{
 \item{\code{"projection",}}{Projection method.}
 \item{\code{"LS",}}{Iterated least squares.}
 \item{\code{"MLE",}}{MLE under a structured covariance tensor.}
 \item{\code{"ar",}}{Stacked vector AR(1) Model.}
}}
}
\value{
a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
}
}
\description{
Estimation function for matrix-valued time series, including the projection method,
the Iterated least squares method and MLE under a structured covariance tensor, as determined by the value of \code{type}.
}
