% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR.RR}
\alias{MAR.RR}
\title{Reduced Rank MAR(1) iterative estimation for matrix-valued time series}
\usage{
MAR.RR(xx, method=c("RRLSE","RRMLE"), LL.init=NULL, RR.init=NULL,Sigl.init=NULL,
Sigr.init=NULL,k1=NULL, k2=NULL, niter=100,tol=1e-6)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times d_2} matrix-valued time series}

\item{method}{character string, specifying the method of the estimation to be used. \describe{
 \item{\code{"RRLSE",}}{Reduced Rank MAR(1) iterative estimation}
 \item{\code{"RRMLE",}}{Reduced Rank MAR(1) iterative estimation with Kronecker covariance structure}
 \item{\code{"VAR",}}{Stacked vector AR(1) Model.}
}}

\item{LL.init}{initial value of \eqn{A_1} in iterations, by default is diagonal matrices}

\item{RR.init}{initial values of \eqn{A_2} in iterations, by default is diagonal matrices}

\item{Sigl.init}{if only \code{method=MLE}, initial value of Sigl, by default is diagonal matrices}

\item{Sigr.init}{if only \code{method=MLE}, initial value of Sigr, by default is diagonal matrices}

\item{k1}{if only \code{method=RRLSE} or \code{method=RRMLE}, rank of \eqn{A_1}, is `NULL` by default}

\item{k2}{if only \code{method=RRLSE} or \code{method=RRMLE}, rank of \eqn{A_2}, is `NULL` by default}

\item{niter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{print.true}{TRUE or FALSE, printe LL and RR}
}
\value{
return a list containing the following:\describe{
\item{\code{LL}}{estimator of LL, a p by p matrix}
\item{\code{RR}}{estimator of RR, a q by q matrix}
\item{\code{Sigl}}{if only \code{method=MLE}, one part of structured covariance matrix \eqn{\Sigma=\Sigma_r \otimes \Sigma_l}}
\item{\code{Sigr}}{if only \code{method=MLE}, one part of structured covariance matrix \eqn{\Sigma=\Sigma_r \otimes \Sigma_l}}
\item{\code{res}}{residual of the MAR(1)}
\item{\code{Sig}}{covariance matrix cov(vec(\eqn{E_t}))}
\item{\code{cov}}{covariance matrix cov(vec(\eqn{A}))}
\item{\code{sd}}{standard error of coefficient matrices \eqn{A}, which is a list with same data structure as \eqn{A}}
\item{\code{dis}}{Frobenius norm difference of last update}
\item{\code{niter}}{number of iterations}
\item{\code{BIC}}{value of information criteria}
}
}
\description{
Reduced Rank MAR(1) iterative estimation for matrix-valued time series,
including the Reduced Rank MAR(1) iterative estimation (RRLSE), and Reduced Rank MAR(1) iterative estimation with Kronecker covariance structure (RRMLE),
which should be specified by the value of \code{method}.
}
\examples{
dim <- c(3,3,3)
xx <- generate.AR(t=500, dim, R=2, P=1, rho=0.5, setting='iid')
est <- MAR.RR(xx, method="RRLSE", k1=1, k2=1)
}
