% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{MAR.SE}
\alias{MAR.SE}
\title{Asymptotic Covariance Matrix of \code{MAR1.otimes}}
\usage{
MAR.SE(xx, B, A, Sigma)
}
\arguments{
\item{xx}{T * p * q matrix-valued time series}

\item{B}{q by q matrix in MAR(1) model}

\item{A}{p by p matrix in MAR(1) model}

\item{Sig}{covariance matrix cov(vec(E_t)) in MAR(1) model}
}
\value{
asmptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of \code{MAR1.otimes} for given A, B and matrix-valued time series xx, see Theory 3 in paper.
}
\examples{
# given T * p * q time series xx
out2=MAR1.LS(xx)
FnormLL=sqrt(sum(out2$LL))
xdim=p;ydim=q
out2Xi=MAR.SE(xx.nm,out2$RR*FnormLL,out2$LL/FnormLL,out2$Sig)
out2SE=sqrt(diag(out2Xi))
SE.A=matrix(out2SE[1:xdim^2],nrow=xdim)
SE.B=t(matrix(out2SE[-(1:xdim^2)],nrow=ydim))
}
