% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{MAR1.RRCC.SE}
\alias{MAR1.RRCC.SE}
\title{Asymptotic Covariance Matrix of Reduced rank MAR(1) with Kronecker covariance structure}
\usage{
MAR1.RRCC.SE((A1,A2,k1,k2,Sigma1,Sigma2,RU1=diag(k1),RV1=diag(k1),RU2=diag(k2),
RV2=diag(k2),mpower=100))
}
\arguments{
\item{A1}{coefficient matrix}

\item{A2}{coefficient matrix}

\item{k1}{rank of A1}

\item{k2}{rank of A2}

\item{Sigma1, Sigma2}{Cov(vec(E_t)) = Sigma1 \otimes Sigma2}

\item{RU1, RV1, RU2, RV2:}{rotation of U1,V1,U2,V2, e.g., new_U1=U1 RU1}

\item{mpower}{truncate vec(X_t) to mpower term, i.e. VMA(mpower). By default is 100.}
}
\value{
a list containing the following:\describe{
\item{\code{Sigma.CC}}{asymptotic covariance matrix of (vec(\hat A1),vec(\hat A2^T))}
\item{\code{Theta1.CC.u}}{asymptotic covariance matrix of vec(\hat U1)}
\item{\code{Theta1.CC.v}}{asymptotic covariance matrix of vec(\hat V1)}
\item{\code{Theta2.CC.u}}{asymptotic covariance matrix of vec(\hat U2)}
\item{\code{Theta2.CC.v}}{asymptotic covariance matrix of vec(\hat V2)}
}
}
\description{
Asymptotic covariance Matrix of Reduced rank MAR(1) Kronecker covariance structure for given a matrix-valued time series xx, see related Theorems in our paper.
}
