% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda.na.iter}
\alias{mfmda.na.iter}
\title{mfmda.na.iter}
\usage{
mfmda.na.iter(Yc, hzero)
}
\arguments{
\item{Yc}{Time Series data for a matrix allowing NaNs}

\item{hzero}{Pre-determined parameter}
}
\value{
The sample version of M matrix
}
\description{
The input data could have NaNs. The estimation approach is iterative.
}
