% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{mfmda.na.vec}
\alias{mfmda.na.vec}
\title{mfmda.na.vec}
\usage{
mfmda.na.vec(Yc, hzero)
}
\arguments{
\item{Yc}{Time Series data for a matrix(dimensions n*p*q), allowing NA input}

\item{hzero}{Pre-scribed parameter h}
}
\value{
The sample version of M matrix
}
\description{
This approach is for the vector-valued estimation with NaNs.
}
