% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{mplot}
\alias{mplot}
\title{Plot Matrix-Valued Time Series}
\usage{
mplot(x)
}
\arguments{
\item{xx}{\eqn{T * m_1 * m_2} tensor-valued time series. Note that the number of mode is 3, where the first mode is time.}
}
\description{
Plot Matrix-Valued time series or Tensor-Valued time series by given mode.
}
\examples{
dim <- c(2,2,2,2)
A <- generateA(dim, R=1)
xx <- generate(dim, T=100)
mplot(xx[,,,1])
}
