% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{mplot}
\alias{mplot}
\title{Plot Matrix-Valued Time Series}
\usage{
mplot(x)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times d_2} matrix-valued time series. Note that the number of mode is 3, where the first mode is time.}
}
\description{
Plot matrix-valued time series or Tensor-Valued time series by given mode.
}
\examples{
dim <- c(3,3,3)
A <- tenAR.A(dim,R=2,P=1,rho)
xx <- tenAR.xx(t=500, A, setting='iid')
mplot(xx[,,1])
}
