% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{PlotNetwork_AB}
\alias{PlotNetwork_AB}
\title{Plot the network graph}
\usage{
PlotNetwork_AB(Ft, iterated_A, iterated_B = iterated_A, labels = use2)
}
\arguments{
\item{Ft}{The estimated factor matrix}

\item{iterated_A}{The left loading matrix}

\item{iterated_B}{The right loading matrix}

\item{labels}{The row labels}
}
\value{
Plot the network graph
#'#'@seealso \code{\link{MAR1.projection}}            # Not decided
}
\description{
Plot the network graph
}
