% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{PM}
\alias{PM}
\title{Permutation matrix PM}
\usage{
PM(m, n)
}
\arguments{
\item{m}{an array of dimentions of matrices \eqn{A_1,A_2,\cdots,A_k}}

\item{n}{length of time}
}
\value{
Permutation matrix PM
#'@seealso \code{\link{trearrange}}
}
\description{
Permutation matrix PM.
}
\examples{
PM(m,n)
}
