% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{projection}
\alias{projection}
\title{Kronecker Product Approximation}
\usage{
projection(A, m1, m2, n1, n2)
}
\arguments{
\item{A}{m by n matrix such that \eqn{m = m1*n1} and \eqn{n = m2*n2}}

\item{m1}{\code{ncol} of A}

\item{m2}{\code{ncol} of B}

\item{n1}{\code{nrow} of A}

\item{n2}{\code{nrow} of B}
}
\value{
a list contaning two estimator (matrix)
}
\description{
Kronecker product approximation used in Projection Method of matrix-value time series.
}
\examples{
A <- matrix(runif(6),ncol=2),
projection(A,3,3,2,2)
}
\seealso{
\code{\link{MAR1.projection}}
}
