% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{TAR1.SE.MLE}
\alias{TAR1.SE.MLE}
\title{Asymptotic Covariance Matrix of \code{TAR1.MLE}}
\usage{
TAR1.SE.MLE(xx, A.true, Sigma)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{A.true}{coefficient matrices in TAR(1) model}

\item{Sigma}{covariance matrix cov(vec(E_t)) in TAR(1) model}
}
\value{
asmptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of \code{TAR1.MLE} for given A tensor-valued time series xx, see related Theorems in our paper.
}
