% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{taxi.sim.AR}
\alias{taxi.sim.AR}
\title{Simulate taxi data by tenAR models}
\usage{
taxi.sim.AR(t, print.tar.coef=FALSE, print.sig=FALSE, seed=123)
}
\arguments{
\item{t}{length of output series.}

\item{print.tar.coef}{print coefficients, default FALSE.}

\item{print.sig}{print covariance matrices, default FALSE.}

\item{seed}{random seed.}
}
\value{
A tensor-valued time series of dimension (5,5,7).
}
\description{
Simulate tensor time series by autoregressive models, using estimated coefficients by taxi data.
}
\examples{
xx = taxi.sim.AR(t=753)
}
\seealso{
\code{\link{taxi.sim.FM}}
}
