% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{taxi.sim.FM}
\alias{taxi.sim.FM}
\title{Simulate taxi data by factor models}
\usage{
taxi.sim.FM(t, print.tar.coef=FALSE, print.loading=FALSE, seed=216)
}
\arguments{
\item{t}{length of output series.}

\item{print.tar.coef}{print coefficients used for simulation, default FALSE.}

\item{print.loading}{print loading matrices used for simulation, default FALSE.}

\item{seed}{random seed.}
}
\value{
A tensor-valued time series of dimension (4,4,3).
}
\description{
Simulate tensor time series by factor models, using estimated coefficients by taxi data.
}
\examples{
xx = taxi.sim.FM(t=252)
}
\seealso{
\code{\link{taxi.sim.AR}}
}
