% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.bic}
\alias{TenAR.bic}
\title{select the number of terms by BIC}
\usage{
TenAR.bic(xx, rmax = 5)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times \cdots \times d_K} tensor-valued time series}
}
\description{
select the number of terms by BIC
}
\examples{
dim <- c(3,3,3)
A <- TenAR.A(dim,R=2,P=1,rho)
xx <- TenAR.xx(t=500, A, setting='iid')
TenAR.bic(xx, rmax=5)
}
