% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{tenAR.predict}
\alias{tenAR.predict}
\title{Model Predictions}
\usage{
tenAR.predict(object, xx, n.head, method="LSE")
}
\arguments{
\item{object}{a fit from TenAR(P) model}

\item{n.head}{number of steps ahead at which to predict.}

\item{method}{method used by rolling forecast}

\item{data}{data to which to apply the prediction.}
}
\value{
predicted value
}
\description{
`tenAR.predict` is a function for predictions from the results of model fitting functions.
The function invokes particular methods which depend on the class of the first argument.
Our function is similar to the usage of classical `predict.ar` in package "stats".
}
\examples{
dim <- c(2,2,2)
A <- tenAR.A(dim,R=2,P=1,rho)
xx <- tenAR.xx(t=500, A, setting='iid')
pred.xx <- tenAR.predict(model, xx, n.head = 5)
}
\seealso{
`predict.ar` or `predict.arima`
}
