% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.PROJ}
\alias{TenAR.PROJ}
\title{Projection estimation for TenAR(p) model}
\usage{
TenAR.PROJ(xx, R, P)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{R}{number of terms}

\item{P}{number of orders}
}
\value{
a list containing the estimation of matrices
}
\description{
projection estimation
}
