% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR}
\alias{TenAR}
\title{Estimation for Autoregressive Model of Tensor-Valued Time Series}
\usage{
TenAR(xx, method)
}
\arguments{
\item{xx}{\eqn{T \times d_1 \times \cdots \times d_K} tensor-valued time series}

\item{R}{number of terms}

\item{P}{number of lag-one orders}

\item{method}{character string, specifying the type of the estimation method to be used. \describe{
 \item{\code{"PROJ",}}{Projection method.}
 \item{\code{"LSE",}}{Iterated least squares.}
 \item{\code{"MLE",}}{MLE under a Kronecker structured covariance matrix.}
 \item{\code{"VAR",}}{Stacked vector AR(1) Model.}
}}
}
\value{
a list containing the following:\describe{
\item{\code{A}}{estimator of coefficient matrices \eqn{A_1,A_2,\cdots,A_K}}
\item{\code{res}}{residual of the model}
\item{\code{Sig}}{covariance matrix cov(vec(E_t))}
\item{\code{niter}}{number of iterations}
}
}
\description{
Estimation function for tensor-valued time series.
Projection method (proj), the Iterated least squares method (LSE), MLE under a Kronecker structured covariance matrix (MLE) and stacked vector AR(1) model (VAR), as determined by the value of \code{type}.
}
\examples{
dim <- c(3,3,3)
A <- TenAR.A(dim,R=2,P=1,rho)
xx <- TenAR.xx(t=500, A, setting='iid')
TenAR(xx, R=1, P=1, method="LSE")
}
