% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenAR.R
\name{TenAR.SE.LSE}
\alias{TenAR.SE.LSE}
\title{Asymptotic Covariance Matrix of LSE in TenAR(1)}
\usage{
TenAR.SE.LSE(xx, A.true, Sigma)
}
\arguments{
\item{xx}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{A.true}{coefficient matrices in TAR(1) model}

\item{Sigma}{covariance matrix cov(vec(E_t)) in TAR(1) model}
}
\value{
asmptotic covariance matrix
}
\description{
Asymptotic covariance Matrix of LSE in TenAR(1) for given a tensor-valued time series xx, see related Theorems in our paper.
}
\examples{
dim <- c(2,2,2)
A <- TenAR.A(dim,R=2,P=1,rho)
xx <- TenAR.xx(t=500, A, setting='iid')
Sigma <- diag(prod(dim))
SIGMA <- TenAR.SE.LSE(xx, A[[1]], Sigma)
}
