% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{tenFM.est}
\alias{tenFM.est}
\title{Tensor factor estimation method}
\usage{
tenFM.est(x,r,method='TIPUP')
}
\arguments{
\item{x}{\eqn{T * m_1 * \cdots * m_K} tensor-valued time series}

\item{r}{input rank of factor tensor}

\item{h0}{lag number}

\item{method}{character string, specifying the type of the estimation method to be used. \describe{
 \item{\code{"TIPUP",}}{TIPUP method}
 \item{\code{"TOPUP",}}{TOPUP method}
}}

\item{iter}{boolean, specifying using an iterative approach or an non-iterative approach}

\item{vmax}{boolean, specifying using varimax rotation on the factor matrix or not}

\item{tol}{tolerance for each iteration}

\item{maxiter}{integer, the max number of iterations allowed}

\item{dim}{dimension of coefficient matrices}
}
\value{
a list containing the following:\describe{
\item{\code{Ft}}{estimated factor matrices of dimension \eqn{T*d_1*d_2 *\cdots *d_K}}
\item{\code{Ft.all}}{Summation of factor matrices over time, dimension \eqn{d_1,d_2,\cdots,d_K} }
\item{\code{Q}}{loading matrices}
\item{\code{x.hat}}{fitted part}
\item{\code{niter}}{number of iterations}
\item{\code{fnorm.resid}}{Frobenius norm of residual}
}
}
\description{
Estimation function for tensor-valued time series factor model.
}
\examples{
Q1 = gen.Q(5,2)
Q2 = gen.Q(5,2)
ar_coef = array(c(0.3,0.5,0.4,0.6),c(2,2))
bF = gen.F(ar_coef,20,1)
x = aperm(gen.X(Q1,Q1,bF),c(3,1,2))
x = as.tensor(x) # Generate x and permute the axes
result <- tenFM.est(x,c(2,2),method='TOPUP')  # Estimate the factor and loadings
Ft <- result$Ft
}
