% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{tenFM.rank}
\alias{tenFM.rank}
\title{Tensor factor model rank determination}
\usage{
tenFM.rank(x,r,rank='BIC',method='TIPUP')
}
\arguments{
\item{x}{\eqn{T \times d_1 \times \cdots \times d_K} dimensional tensor-valued time series}

\item{r}{input rank of factor matrix}

\item{h0}{lag number}

\item{rank}{character string, specifying the type of the rank determination method to be used. \describe{
 \item{\code{"BIC",}}{Bayesian information criterion}
 \item{\code{"ER",}}{Eigen ratio method}
}}

\item{method}{character string, specifying the type of the factor estimation method to be used. \describe{
 \item{\code{"TIPUP",}}{TIPUP method}
 \item{\code{"TOPUP",}}{TOPUP method}
}}

\item{inputr}{boolean, if TRUE, use input rank for each iteration; if FLASE, update the rank r in each iteration}

\item{iter}{boolean, specifying using an iterative approach or an non-iterative approach}

\item{penalty}{takes value in {1,2,3,4,5}, decide which penalty function to use. \describe{
 \item{}{When \code{rank}= 'BIC':}
 \item{}{if \code{penalty}=1, \eqn{g_1= \frac{h_0 d^{2-2\nu}}{T}\log(\frac{dT}{d+T})}}
 \item{}{if \code{penalty}=2, \eqn{g_2= h_0 d^{2-2\nu}(\frac{1}{T}+\frac{1}{d})\log(\frac{dT}{d+T})}}
 \item{}{if \code{penalty}=3, \eqn{g_3= \frac{h_0 d^{2-2\nu}}{T} \log(\min{d,T})}}
 \item{}{if \code{penalty}=4, \eqn{g_4= h_0 d^{2-2\nu}(\frac{1}{T}+\frac{1}{d})\log(\min{d,T})}}
 \item{}{if \code{penalty}=5, \eqn{g_5= h_0 d^{2-2\nu}(\frac{1}{T}+\frac{1}{d})\log(\min{d_k,T})}}
 \item{}{When \code{rank}= 'ER':}
 \item{}{if \code{penalty}=1, \eqn{h_1= c_0 h_0}}
 \item{}{if \code{penalty}=2, \eqn{h_2= \frac{h_0 d^2}{T^2}}}
 \item{}{if \code{penalty}=3, \eqn{h_3= \frac{h_0 d^2}{T^2 d_k^2}}}
 \item{}{if \code{penalty}=4, \eqn{h_4= \frac{h_0 d^2}{T^2 d_k^2} + \frac{h_0 d_k^2}{T^2}}}
 \item{}{if \code{penalty}=5, \eqn{h_5= \frac{h_0 d^2}{T^2 d_k^2} + \frac{h_0 dd_k^2}{T^2}}}
}}

\item{delta1}{weakest factor strength, a tuning parameter used for BIC method only}

\item{tol}{tolerance for each iteration}

\item{maxiter}{integer, the max number of iterations allowed}
}
\value{
estimated number of factors \eqn{n_1,n_2,\cdots,n_k}
}
\description{
Function for rank determination of the tensor factor model
}
\examples{
dims <- c(4,5,6) # dimensions of tensor time series
t <- 50
r <- c(2,2,2)  # dimensions of factor series
x = tenFM.sim(dims,array(c(runif(prod(r)),min=0.5,max=1),r),t)  # generate t*dims tensor
Et <- array(rnorm(t*prod(dims)),c(t,dims))  # add some iid noise to the tensor time series
x = x + Et
rank <- tenFM.rank(x,r,rank='BIC',method='TIPUP')  # Estimate the rank
}
