% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFM.R
\name{tensor.ratio}
\alias{tensor.ratio}
\title{Eigenvalue ratio estimators for determining the numbers of factors}
\usage{
tensor.ratio(reigen, p1, p2, n)
}
\arguments{
\item{reigen}{list of eigenvalues}

\item{p1}{p1}

\item{p2}{p2}

\item{n}{n}

\item{h0}{lag number}
}
\value{
factor.p1: Estimated number of factors
}
\description{
Eigenvalue ratio estimators for determining the numbers of factors
}
