% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFA.R
\docType{package}
\name{timeFA}
\alias{timeFA}
\title{timeFA: factor and autoregressive models for high dimensional time series.}
\description{
The timefa package provides two categories of models:
factor models and autoregressive models.
}
\section{factor model functions}{

The functions ...
}

\section{autoregressive model functions}{

The functions ...
}

