% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{vts}
\alias{vts}
\title{vector time series}
\usage{
vts(x, h0, r)
}
\arguments{
\item{x}{a n*d matrix}

\item{h0}{Pre-scribed parameter h}

\item{r}{First r eigenvectors}
}
\value{
a list containing the following:\describe{
\item{\code{M}}{.}
\item{\code{Q}}{The eigenvectors of matrix M}
}
}
\description{
vector time series
}
